refactor(detectOs): create a type userOS
