+++ types/userOS.ts
@@ -1,0 +1,1 @@
+export type UserOS = 'MAC' | 'WIN' | 'LINUX' | 'UNKNOWN';
